Guillaume wrote: "In NSIndexPath, there is a NSIndexPath UI Kit Additions.  It's not listed in AppKido."  The fix was to tweak the logic in AKCocoaBehaviorDocParser for deciding whether a file contains the class's "main" reference.  I was checking for "...Additions Reference" in the title, and the doc Guillaume referred to is called "NSIndexPath UIKit Additions" (without the word "Reference").
When populating "classes with delegates/datasources", added a check for classes that have *properties* named "delegate"/"datasource", and also whether protocol ThisClassDelegate/ThisClassDataSource exists.

UITextField's delegate methods are documented in the doc for UITextFieldDelegate, and not in the class's doc file, they weren't showing up in the "Delegate Methods" doc list.  Fixed by adding logic to _addExtraDelegateMethodsTo: to see if there exists a protocol named ThisClassDelegate.

The delegate methods doc list for NSApplication showed "application:printFiles: <(null)>".  Problem was a missing call to setOwningBehavior: when adding deprecated delegate method.  Changed AKMemberNode to take owningBehavior in its designated initializer, and got rid of setOwningBehavior:, to avoid this problem in the future.


git-svn-id: http://svn.appkido.com/svn/appkido/trunk@131 aeb25c08-0953-0410-bd6c-f302445d6ae0