Fixed bug reported by Jeff Laing: NSTask, which doesn't exist on the iPhone, was showing up in AppKiDo-for-iPhone.  The problem was that AKDocSetIndex was looking in iPhoneSimulator.platform for headers, when it should have been looking in iPhoneOS.platform.
While fixing this, took the opportunity to change the logic that was deriving header directories from a list of header file paths and then iterating through those directories to find the files it just had a list of, along with other header files we might not want.  Now simply iterates through the files we had in the first place.  Compared before-and-after XML dumps and found that in WebKit, this got rid of 104 DOM classes and protocols that I'd been unnecessarily creating nodes for in the database (possibly would have been more if I'd included more frameworks in my prefs).

git-svn-id: http://svn.appkido.com/svn/appkido/trunk@161 aeb25c08-0953-0410-bd6c-f302445d6ae0