Fixed (at least for me) the bug reported by Gerriet where NSFileVersion wasn't appearing.
Explanation: It is possible for docsets with the same name (i.e., for the same SDK) to appear in both /Library/Developer/Shared/Documentation/DocSets and /Library/Developer/Documentation/DocSets. In my case the one in Shared was newer, so I changed things so that it "wins". Hopefully I can count on this being correct in the general case.