Fixed a parsing bug causing UIKit rather than Foundation to be set as the owning framework for NSObject in AppKiDo-for-iPhone. So when you selected NSObject, the displayed description was "UIKit class" rather than "Foundation class". If you then selected the "General" subtopic you'd see "Overview" and "Overview [Foundation Additions]" when it should have been "Overview" and "Overview [UIKit Additions]".
My error was to assume that curly braces inside an @interface declaration are always an ivar declaration section, and therefore it must be a class being declared and not a category. That was not correct, because an enum declaration can also be there. The file I was thus incorrectly parsing was /Applications/Xcode-4.5.2.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS6.0.sdk/System/Library/Frameworks/UIKit.framework/Headers/UIAccessibility.h.